requirements.txt
Django==2.1.1
pytz==2018.5
djangorestframework